requirements.txt
Django>=4
djangorestframework
channels
channels-redis
pytest-asyncio
pytest-django
Pillow
djangorestframework-simplejwt
psycopg2-binary